Restoring playlists works as well as it's going to.
Because playlists are not stored by uid, and the dashboards
within them are not referenced by uids, we have to search by
name when updating existing playlists, and we have to
reconstruct the contents of the playlist based on dashboard names
as we restore the playlist.  This works, but is prone to error.